+++ src/antares/study/version/create_app/__init__.py
@@ -1,6 +1,5 @@
 import dataclasses
 import datetime
-import typing as t
 import zipfile
 from pathlib import Path